Replaced IntegrationDistance->MagDepDistance [demos]
