+++ openquake/calculators/classical.py
@@ -34,7 +34,7 @@ from openquake.baselib.python3compat import encode
 from openquake.baselib.general import (
     AccumDict, DictArray, block_splitter, groupby, humansize, get_array_nbytes)
 from openquake.hazardlib.contexts import ContextMaker, get_effect
-from openquake.hazardlib.calc.filters import split_sources, getdefault
+from openquake.hazardlib.calc.filters import split_sources
 from openquake.hazardlib.calc.hazard_curve import classical
 from openquake.hazardlib.probability_map import ProbabilityMap
 from openquake.commonlib import calc, util, logs, readinput
@@ -324,6 +324,7 @@ class ClassicalCalculator(base.HazardCalculator):
         tectonic region type.
         """
         oq = self.oqparam
+        psd = oq.pointsource_distance
         if oq.hazard_calculation_id and not oq.compare_with_classical:
             with util.read(self.oqparam.hazard_calculation_id) as parent:
                 self.full_lt = parent['full_lt']
@@ -334,37 +335,30 @@ class ClassicalCalculator(base.HazardCalculator):
         if len(mags) == 0:  # everything was discarded
             raise RuntimeError('All sources were discarded!?')
         gsims_by_trt = self.full_lt.get_gsims_by_trt()
-        if oq.pointsource_distance is not None:
-            for trt in gsims_by_trt:
-                oq.pointsource_distance[trt] = getdefault(
-                    oq.pointsource_distance, trt)
         mags_by_trt = {}
         for trt in mags:
             mags_by_trt[trt] = mags[trt][()]
-        imts_with_period = [imt for imt in oq.imtls
-                            if imt == 'PGA' or imt.startswith('SA')]
-        imts_ok = len(imts_with_period) == len(oq.imtls)
-        if (imts_ok and oq.pointsource_distance and
-                oq.pointsource_distance.suggested()) or (
-                    imts_ok and oq.minimum_intensity):
-            aw, self.psd = get_effect(
-                mags_by_trt, self.sitecol.one(), gsims_by_trt, oq)
-            dic = {trt: [(float(mag), int(dst))
-                         for mag, dst in self.psd[trt].items()]
-                   for trt in self.psd if trt != 'default'}
-            logging.info('pointsource_distance=\n%s', pprint.pformat(dic))
-            if len(vars(aw)) > 1:  # more than _extra
-                self.datastore['effect_by_mag_dst'] = aw
-        elif oq.pointsource_distance:
-            self.psd = oq.pointsource_distance.interp(mags_by_trt)
-            for trt, dic in self.psd.items():
+        oq.maximum_distance.interp(mags_by_trt)
+        if psd is not None:
+            psd.interp(mags_by_trt)
+            for trt, dic in psd.ddic.items():
                 # the sum is zero for {'default': [(1, 0), (10, 0)]}
                 if sum(dic.values()):
                     it = list(dic.items())
                     md = '%s->%d ... %s->%d' % (it[0] + it[-1])
                     logging.info('ps_dist %s: %s', trt, md)
-        else:
-            self.psd = {}
+        imts_with_period = [imt for imt in oq.imtls
+                            if imt == 'PGA' or imt.startswith('SA')]
+        imts_ok = len(imts_with_period) == len(oq.imtls)
+        if (imts_ok and psd and psd.suggested() or
+                (imts_ok and oq.minimum_intensity)):
+            aw = get_effect(mags_by_trt, self.sitecol.one(), gsims_by_trt, oq)
+            dic = {trt: [(float(mag), int(dst))
+                         for mag, dst in psd[trt].items()]
+                   for trt in psd if trt != 'default'}
+            logging.info('pointsource_distance=\n%s', pprint.pformat(dic))
+            if len(vars(aw)) > 1:  # more than _extra
+                self.datastore['effect_by_mag_dst'] = aw
         smap = parallel.Starmap(classical, h5=self.datastore.hdf5,
                                 num_cores=oq.num_cores)
         self.submit_tasks(smap)
@@ -400,8 +394,8 @@ class ClassicalCalculator(base.HazardCalculator):
             int(self.numrups), self.totrups))
         logging.info('Effective number of sites per rupture: %d',
                      numsites / self.numrups)
-        if self.psd:
-            psdist = max(max(self.psd[trt].values()) for trt in self.psd)
+        if psd:
+            psdist = max(max(psd.ddic[trt].values()) for trt in psd.ddic)
             if psdist and self.maxradius >= psdist / 2:
                 logging.warning('The pointsource_distance of %d km is too '
                                 'small compared to a maxradius of %d km',
@@ -449,7 +443,7 @@ class ClassicalCalculator(base.HazardCalculator):
             truncation_level=oq.truncation_level, imtls=oq.imtls,
             filter_distance=oq.filter_distance, reqv=oq.get_reqv(),
             maximum_distance=oq.maximum_distance,
-            pointsource_distance=self.psd,
+            pointsource_distance=oq.pointsource_distance,
             point_rupture_bins=oq.point_rupture_bins,
             shift_hypo=oq.shift_hypo, max_weight=max_weight,
             collapse_level=oq.collapse_level,
@@ -478,11 +472,8 @@ class ClassicalCalculator(base.HazardCalculator):
 
             w = sum(srcweight(src) for src in sg)
             logging.info('TRT = %s', sg.trt)
-            if oq.maximum_distance.magdist:
-                it = sorted(oq.maximum_distance.magdist[sg.trt].items())
-                md = '%s->%d ... %s->%d' % (it[0] + it[-1])
-            else:
-                md = oq.maximum_distance(sg.trt)
+            it = sorted(oq.maximum_distance.ddic[sg.trt].items())
+            md = '%s->%d ... %s->%d' % (it[0] + it[-1])
             logging.info('max_dist={}, gsims={}, weight={:_d}, blocks={}'.
                          format(md, len(gsims), int(w), nb))
