+++ openquake/hazardlib/calc/filters.py
@@ -17,6 +17,7 @@
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
 import os
 import re
+import ast
 import sys
 import time
 import logging
@@ -69,29 +70,106 @@ def getdefault(dic_with_default, key):
         return dic_with_default['default']
 
 
-class IntegrationDistance(collections.abc.Mapping):
+def unique_sorted(items):
     """
-    Pickleable object wrapping a dictionary of integration distances per
-    tectonic region type. Here is an example using 'default'
-    as tectonic region type, so that the same values will be used for all
-    tectonic region types:
+    Check that the items are unique and sorted
+    """
+    if len(set(items)) < len(items):
+        raise ValueError('Found duplicates in %s' % items)
+    elif items != sorted(items):
+        raise ValueError('%s is not ordered' % items)
+    return items
 
-    >>> maxdist = IntegrationDistance({'default': 400})
-    >>> maxdist('Some TRT')
-    400
-    >>> maxdist('Some TRT', mag=2.5)
-    400
+
+# used for the maximum distance parameter in the job.ini file
+def floatdict(value):
     """
-    def __init__(self, dic):
-        self.dic = dic  # TRT -> float
-        self.magdist = {}  # TRT -> (magnitudes, distances), set by the engine
+    :param value:
+        input string corresponding to a literal Python number or dictionary
+    :returns:
+        a Python dictionary key -> number
+
+    >>> floatdict("200")
+    {'default': 200}
+
+    >>> text = "{'active shallow crust': 250., 'default': 200}"
+    >>> sorted(floatdict(text).items())
+    [('active shallow crust', 250.0), ('default', 200)]
+    """
+    value = ast.literal_eval(value)
+    if isinstance(value, (int, float, list)):
+        return {'default': value}
+    dic = {'default': max(value.values())}
+    dic.update(value)
+    return dic
+
+
+class MagDepDistance(dict):
+    """
+    A dictionary trt -> [(mag, dist), ...]
+    """
+    @classmethod
+    def new(cls, value):
+        """
+        :param value: string to be converted
+        :returns: MagDepDistance dictionary
+
+        >>> md = MagDepDistance.new('50')
+        >>> md
+        {'default': [(1, 50), (10, 50)]}
+        >>> md.max()
+        {'default': 50}
+        >>> md.interp(dict(default=['5.0', '5.1', '5.2']))
+        {'default': {'5.0': 50.0, '5.1': 50.0, '5.2': 50.0}}
+        """
+        items_by_trt = floatdict(value.replace('?', '-1'))
+        self = cls()
+        for trt, items in items_by_trt.items():
+            if isinstance(items, list):
+                self[trt] = unique_sorted(items)
+                for mag, dist in self[trt]:
+                    if mag < 1 or mag > 10:
+                        raise ValueError('Invalid magnitude %s' % mag)
+            else:  # assume scalar distance
+                assert items == -1 or items >= 0, items
+                self[trt] = [(1, items), (10, items)]
+        return self
+
+    def interp(self, mags_by_trt):
+        """
+        :param mags_by_trt: a dictionary trt -> magnitudes as strings
+        :returns: a dictionary trt->mag->dist
+        """
+        ddic = {}
+        for trt, mags in mags_by_trt.items():
+            xs, ys = zip(*getdefault(self, trt))
+            if len(mags) == 1:
+                ms = [numpy.float64(mags)]
+            else:
+                ms = numpy.float64(mags)
+            dists = numpy.interp(ms, xs, ys)
+            ddic[trt] = dict(zip(mags, dists))
+        self.ddic = ddic
 
     def __call__(self, trt, mag=None):
-        if mag and trt in self.magdist:
-            return self.magdist[trt]['%.2f' % mag]
-        elif not self.dic:
+        if mag is None or not hasattr(self, 'ddic'):
+            return getdefault(self, trt)[-1][1]
+        elif mag and trt in self.ddic:
+            return self.ddic[trt]['%.2f' % mag]
+        else:
             return MAX_DISTANCE
-        return getdefault(self.dic, trt)
+
+    def max(self):
+        """
+        :returns: a dictionary trt -> maxdist
+        """
+        return {trt: self[trt][-1][1] for trt in self}
+
+    def suggested(self):
+        """
+        :returns: True if there is a ? for any TRT
+        """
+        return any(self[trt][-1][1] == -1 for trt in self)
 
     def get_bounding_box(self, lon, lat, trt=None, mag=None):
         """
@@ -105,7 +183,7 @@ class IntegrationDistance(collections.abc.Mapping):
         :returns: min_lon, min_lat, max_lon, max_lat
         """
         if trt is None:  # take the greatest integration distance
-            maxdist = max(self(trt, mag) for trt in self.dic)
+            maxdist = max(self(trt, mag) for trt in self)
         else:  # get the integration distance for the given TRT
             maxdist = self(trt, mag)
         a1 = min(maxdist * KM_TO_DEGREES, 90)
@@ -132,28 +210,12 @@ class IntegrationDistance(collections.abc.Mapping):
         """
         return .01 + numpy.arange(nbins) * self(trt) / (nbins - 1)
 
-    def __getstate__(self):
-        # otherwise is not pickleable due to .piecewise
-        return dict(dic=self.dic, magdist=self.magdist)
-
-    def __getitem__(self, trt):
-        return self(trt)
-
-    def __iter__(self):
-        return iter(self.dic)
-
-    def __len__(self):
-        return len(self.dic)
-
     def __toh5__(self):
-        dic = {trt: numpy.array(dist) for trt, dist in self.dic.items()}
+        dic = {trt: numpy.array(dist) for trt, dist in self.items()}
         return dic, {}
 
     def __fromh5__(self, dic, attrs):
         self.__init__({trt: dic[trt][()] for trt in dic})
-
-    def __repr__(self):
-        return repr(self.dic)
 
 
 def split_sources(srcs):
@@ -232,7 +294,7 @@ class SourceFilter(object):
             integration_distance = {}
         self.filename = filename
         self.integration_distance = (
-            IntegrationDistance(integration_distance)
+            MagDepDistance(integration_distance)
             if isinstance(integration_distance, dict)
             else integration_distance)
         if not filename:  # keep the sitecol in memory
