+++ openquake/hazardlib/valid.py
@@ -32,7 +32,7 @@ from openquake.baselib import hdf5
 from openquake.hazardlib import imt, scalerel, gsim, pmf, site
 from openquake.hazardlib.gsim.base import registry, gsim_aliases
 from openquake.hazardlib.calc import disagg
-from openquake.hazardlib.calc.filters import IntegrationDistance
+from openquake.hazardlib.calc.filters import MagDepDistance, floatdict
 
 PRECISION = pmf.PRECISION
 
@@ -875,109 +875,6 @@ def dictionary(value):
         if has_logscale:
             dic[key] = list(logscale(*val[1:]))
     return dic
-
-
-# used for the maximum distance parameter in the job.ini file
-def floatdict(value):
-    """
-    :param value:
-        input string corresponding to a literal Python number or dictionary
-    :returns:
-        a Python dictionary key -> number
-
-    >>> floatdict("200")
-    {'default': 200}
-
-    >>> text = "{'active shallow crust': 250., 'default': 200}"
-    >>> sorted(floatdict(text).items())
-    [('active shallow crust', 250.0), ('default', 200)]
-    """
-    value = ast.literal_eval(value)
-    if isinstance(value, (int, float, list)):
-        return {'default': value}
-    dic = {'default': max(value.values())}
-    dic.update(value)
-    return dic
-
-
-def maximum_distance(value):
-    """
-    :param value:
-        input string corresponding to a valid maximum distance
-    :returns:
-        a IntegrationDistance mapping
-    """
-    return IntegrationDistance(MagDist.new(value).max())
-
-
-def unique_sorted(items):
-    """
-    Check that the items are unique and sorted
-    """
-    if len(set(items)) < len(items):
-        raise ValueError('Found duplicates in %s' % items)
-    elif items != sorted(items):
-        raise ValueError('%s is not ordered' % items)
-    return items
-
-
-class MagDist(dict):
-    """
-    A dictionary trt -> [(mag, dist), ...]
-    """
-    @classmethod
-    def new(cls, value):
-        """
-        :param value: string to be converted
-        :returns: MagDist dictionary
-
-        >>> md = MagDist.new('50')
-        >>> md
-        {'default': [(1, 50), (10, 50)]}
-        >>> md.max()
-        {'default': 50}
-        >>> md.interp(dict(default=['5.0', '5.1', '5.2']))
-        {'default': {'5.0': 50.0, '5.1': 50.0, '5.2': 50.0}}
-        """
-        items_by_trt = floatdict(value.replace('?', '-1'))
-        self = cls()
-        for trt, items in items_by_trt.items():
-            if isinstance(items, list):
-                self[trt] = unique_sorted(items)
-                for mag, dist in self[trt]:
-                    magnitude(mag)  # check valid magnitude
-            else:  # assume scalar distance
-                assert items == -1 or items >= 0, items
-                self[trt] = [(1, items), (10, items)]
-        return self
-
-    def interp(self, mags_by_trt):
-        """
-        :param mags_by_trt: a dictionary trt -> magnitudes as strings
-        :returns: a dictionary trt->mag->dist
-        """
-        dic = {}
-        for trt, mags in mags_by_trt.items():
-            xs, ys = zip(*self[trt])
-            if len(mags) == 1:
-                ms = [numpy.float64(mags)]
-            else:
-                ms = numpy.float64(mags)
-            dists = numpy.interp(ms, xs, ys)
-            dic[trt] = dict(zip(mags, dists))
-        return dic
-
-    def max(self):
-        """
-        :returns: a dictionary trt -> maxdist
-        """
-        return {trt: self[trt][-1][1] for trt in self}
-
-    def suggested(self):
-        """
-        :returns: True if there is a * for any TRT
-        """
-        return any(self[trt][-1][1] == -1 for trt in self)
 
 
 # ########################### SOURCES/RUPTURES ############################# #
